- Fix test check fully-qualified airline can register a second airline

+++ project4-ethereum-flight-surety-dapp/contracts/data/DataInsurerController.sol
@@ -27,7 +27,7 @@ abstract contract DataInsurerController is PayableContract, DataOperationalContr
         mapping(address => bool) approvers;
     }
 
-    uint16 fullyQualifiedInsurersCtr;
+    uint fullyQualifiedInsurersCtr;
     mapping(address => InsurerProfile) private insurers;
 
     /**
